Did quite a bit.
Finished adding the images and updating the map array.
Also started the actual function to check if the player can move in that
direction or not.
. . . it would have helped if I actually instantiated the coordinate
mapping function in the first place. . .
Anywho, it works very well except for the fact that if it hits a limit (
for example you go down all the way) and you keep going down, it will
continue to let you do so, but not any other direction.