Added some basic attack addition functions.
Need to add in the various functions for HP and attack (that have
limits) to check the amount being added and make sure that the existing
plus the amount isn’t greater than the limit. Should be simple to
implement. If that’s the case, then it’ll always be the limit.