Added shop button and fixed character positions
Did not notice this, but originally the char pos were at a fixed
position to how my browser was positioned, so when resized. it was
screwed up. This has been changed to fix itself to whatever screen size
there is.

Added a button over the shop, clickable to bring up the shop menu.